Implemented Saving an Event

+++ Trackify/ViewModels/EventFormViewModel.cs
@@ -8,7 +8,7 @@ namespace Trackify.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Category { get; set; }
+        public byte Category { get; set; }
         public IEnumerable<Category> Categories { get; set; }
     }
 }
